requirements.txt
django==3.2.25
boto3>=1.26.0
botocore>=1.29.0
python-dotenv==0.21.1
